Create Grain128-PLE
This script executes a modification to the Grain-128 AEAD v2 stream cipher (finalist in the Lightweight Cryptography Competition by NIST) such that it can be implemented as a generic encryption scheme at the physical layer (after channel coding, prior to signal modulation). This Grain128-PLE scheme can be interpreted as a replacement of the insecure (and outdated) A5/1 stream cipher that was implemented at the physical layer to achieve data confidentiality for GSM communications systems.